refactor: split clinician profile page under 300-line limit
Extract Practice Schedule card into ClinicianPracticeSchedule component
Extract Unavailability card into ClinicianUnavailabilityCard component
Remove max-lines eslint-disable (file now 295 lines vs 300 limit)
Includes previous: switch lookup, counter, null-guard fixes
