+++ src/app/profile/clinician.tsx
@@ -1,11 +1,10 @@
 'use client';
-/* eslint-disable react/jsx-max-depth */
 
 import { LoadingSpinnerIcon } from '@/components/icons';
+import ClinicianPracticeSchedule from '@/components/profile/clinician-practice-schedule';
+import ClinicianUnavailabilityCard from '@/components/profile/clinician-unavailability-card';
 import InformationDetail from '@/components/profile/information-detail';
 import Settings from '@/components/profile/settings';
-import Tags from '@/components/profile/tags';
-import MarkUnavailabilityButton from '@/components/schedule/mark-unavailability';
 import {
   Drawer,
   DrawerContent,
@@ -24,7 +23,7 @@ import { findAge, generateAvatarPlaceholder, mapAddress } from '@/utils/helper';
 import { useQuery } from '@tanstack/react-query';
 import { format } from 'date-fns';
 import { Practitioner, PractitionerRole } from 'fhir/r4';
-import Image from 'next/image';
+
 import { useRouter } from 'next/navigation';
 import { useEffect, useState } from 'react';
 import { toast } from 'react-toastify';
@@ -148,7 +147,8 @@ export default function Clinician({ fhirId }: Props) {
     > = {};
 
     activeFirms.forEach(role => {
-      const organizationName = role?.organizationData.name || '';
+      if (!role) return;
+      const organizationName = role.organizationData?.name || '';
 
       if (Array.isArray(role.availableTime)) {
         // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -256,96 +256,12 @@ export default function Clinician({ fhirId }: Props) {
         />
       )}
 
-      {/* display practitioner's availability schedules */}
-      <div className='mt-4 flex w-full flex-col items-center justify-center rounded-[16px] border-0 bg-[#F9F9F9] p-4'>
-        {/* Practice Schedule section title - moved to top with styling to match other sections */}
-        <div className='flex w-full items-center justify-between'>
-          <div className='flex w-1/2 items-center'>
-            <Image
-              src={'/icons/calendar-profile.svg'}
-              width={30}
-              height={30}
-              alt='calendar-icon'
-              className='pr-[13px]'
-            />
-            <p className='flex-grow text-start text-[10px] font-normal text-[#2C2F35] opacity-40'>
-              Practice Schedule
-            </p>
-          </div>
-          <div className='flex w-1/2 items-center justify-end'>
-            <button
-              onClick={handleOpenDrawer}
-              className='cursor-pointer transition-all duration-200 hover:brightness-90'
-            >
-              <div className='bg-secondary w-[100px] rounded-full p-[7px]'>
-                <p className='text-[10px] text-white'>Edit Schedule</p>
-              </div>
-            </button>
-          </div>
-        </div>
+      <ClinicianPracticeSchedule
+        groupedByFirmAndDay={groupedByFirmAndDay}
+        onEditSchedule={handleOpenDrawer}
+      />
 
-        {/* Availability content with border divider */}
-        <div className='mt-2 flex w-full flex-col border-t border-[#E3E3E3]'>
-          {Object.keys(groupedByFirmAndDay).map(firm => {
-            const availability = groupedByFirmAndDay[firm].availability;
-            return (
-              <div key={firm}>
-                <div className='mb-1 text-start text-[14px] font-bold'>
-                  {firm}
-                </div>
-                {Object.keys(availability).map(day => {
-                  const timeRanges = availability[day] || [];
-                  const tags = timeRanges.map(
-                    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-                    (timeRange: any) =>
-                      `${day}: ${timeRange.fromTime} - ${timeRange.toTime}`
-                  );
-
-                  return (
-                    <div
-                      key={`${firm}-${day}`}
-                      className='mb-1 flex w-full flex-wrap gap-[10px]'
-                    >
-                      <Tags tags={tags} />
-                    </div>
-                  );
-                })}
-              </div>
-            );
-          })}
-        </div>
-      </div>
-
-      {/* Current Unavailability section - separate section with same style as Practice Schedule */}
-      <div className='mt-4 flex w-full flex-col items-center justify-center rounded-[16px] border-0 bg-[#F9F9F9] p-4'>
-        <div className='flex w-full items-center justify-between'>
-          <div className='flex w-1/2 items-center'>
-            <Image
-              src={'/icons/calendar-profile.svg'}
-              width={30}
-              height={30}
-              alt='calendar-icon'
-              className='pr-[13px]'
-            />
-            <p className='flex-grow text-start text-[10px] font-normal text-[#2C2F35] opacity-40'>
-              Current Unavailability
-            </p>
-          </div>
-          <div className='flex w-1/2 items-center justify-end'>
-            <MarkUnavailabilityButton
-              triggerClassName='cursor-pointer hover:brightness-90 transition-all duration-200'
-              buttonText='Mark Away'
-            />
-          </div>
-        </div>
-
-        {/* Unavailability content with border divider */}
-        <div className='mt-2 flex w-full flex-col border-t border-[#E3E3E3]'>
-          <div className='py-2 text-center text-[14px] text-[#2C2F35]'>
-            No Unavailability
-          </div>
-        </div>
-      </div>
+      <ClinicianUnavailabilityCard />
 
       <Settings menus={settingMenus} />
