+++ src/utils/role-fhir.ts
@@ -1,14 +1,16 @@
 import { Roles } from '@/constants/roles';
 
-const ROLE_TO_FHIR_RESOURCE: Record<string, string> = {
-  [Roles.Patient]: 'Patient',
-  [Roles.Practitioner]: 'Practitioner',
-  [Roles.ClinicAdmin]: 'Person'
-};
-
 /**
  *
  */
 export function roleToFhirResource(role: string): string {
-  return ROLE_TO_FHIR_RESOURCE[role] || 'Patient';
+  switch (role) {
+    case Roles.Practitioner:
+      return 'Practitioner';
+    case Roles.ClinicAdmin:
+      return 'Person';
+    case Roles.Patient:
+    default:
+      return 'Patient';
+  }
 }
